requirements-eval.txt
arize-phoenix
arize-phoenix-otel
arize-phoenix-evals
openinference-instrumentation-openai
opentelemetry-api
opentelemetry-sdk
sqlparse
pyyaml
datacompy
scikit-learn
